+++ source/Public/New-CompositeResourceModule.ps1
@@ -80,7 +80,7 @@ function New-CompositeResourceModule
                 }
                 AppCredentials = @(
                     @{
-                        Workload       = 'String | Required | Name of the Workload for which this credential will be used | AzureAD / AzureDevOps / Azure / Defender / Exchange / Fabric / Intune / Office365 / OneDrive / Planner / PowerPlatform / SecurityCompliance / Sentinel / SharePoint / Teams'
+                        Workload       = 'String | Required | Name of the Workload for which this credential will be used | AzureAD / AzureDevOps / Azure / Commerce / Defender / Exchange / Fabric / Intune / Office365 / OneDrive / Planner / PowerPlatform / SecurityCompliance / Sentinel / SharePoint / Teams'
                         ApplicationId  = 'Guid | Required | The GUID of the Entra ID Service Principal'
                         CertThumbprint = 'String | Required | The Certificate Thumbprint of the certificate used for authentication'
                     }
@@ -157,6 +157,11 @@ function New-CompositeResourceModule
                     { $_.StartsWith('Azure') }
                     {
                         $resourceWorkload = 'Azure'
+                        $customResourceName = $shortResourceName -replace "^$resourceWorkload"
+                    }
+                    { $_.StartsWith('Commerce') }
+                    {
+                        $resourceWorkload = 'Commerce'
                         $customResourceName = $shortResourceName -replace "^$resourceWorkload"
                     }
                     { $_.StartsWith('Defender') }